ENH: renamed 'GuessParams' to 'ParamModel'
