+++ pyxrf/model/fit_spectrum.py
@@ -28,7 +28,7 @@ from skbeam.core.fitting.xrf_model import (ModelSpectrum, update_parameter_dict,
                                            # linear_spectrum_fitting,
                                            register_strategy, TRANSITIONS_LOOKUP)
 from skbeam.fluorescence import XrfElement as Element
-from .guessparam import (calculate_profile, fit_strategy_list,
+from .parameters import (calculate_profile, fit_strategy_list,
                          trim_escape_peak, define_range, get_energy,
                          get_Z, PreFitStatus, ElementController,
                          update_param_from_element)
